+++ NEWS.md
@@ -23,6 +23,7 @@
 
 ## Changes to functions
 
+- The `print()`-method for `model_parameters()` gets a `select`-argument, to print only selected columns of the parameters table.
 - `model_parameters()` for mixed models gets a `summary_random`-argument to additionally print information about the random effects.
 - `model_parameters()` for *lme*-models now also accepts the `df_method`-argument.
 - The Intercept-parameter in `model_parameters()` for meta-analysis models was renamed to `"Overall"`.